Login and registration working, uses different table then FB login
Got login and registration working, I use a simpler table then the FB
login does so you much rebuild the database. I was planing on
refactoring the FB login to work with the new DB but failed, if you
want to use both FB and manual a new DB with both tables will need to
be made.